app/requirements.txt
asgiref==3.3.1
boto3==1.11.12
botocore==1.14.17
Django==3.1.6
django-sendgrid-v5==0.9.1
django-storages==1.9.1
django-tinymce==3.2.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
pytz==2021.1
s3transfer==0.3.4
six==1.10.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.25.11
whitenoise==5.2.0
